Updated UltimateBravery.java
added all of the code it needs to produce images on the screen. Will need to have the xml file optimized for layout and sizing. This is horrible to look at, but it was the only way to call image names, as I couldn't get a way to use a string and cast it to a part of an argument, i.e. "R.drawable.riven" from "riven"