Aaandd... Build Guides. (Description)
BuildDatabase.java ready for items! Just add the strings to the correct
locations based on starting items, rush, or as needed. The first slot in
"start" is for a trinket.

There are different methods for each possible role that is passed. Even
if a role is a little silly, if you can find a reasonable build guide,
go for it. Otherwise just leave it be with the constructor and no
setting functions.